Allow custom object holder references to be registered
